+++ src/app/modules/lux-error/lux-http-error/lux-http-error-interceptor.ts
@@ -1,10 +1,10 @@
 import { HttpErrorResponse, HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { tap } from 'rxjs/operators';
 import { Observable, ReplaySubject } from 'rxjs';
+import { tap } from 'rxjs/operators';
 
 @Injectable({
-  providedIn: "root"
+  providedIn: 'root'
 })
 export class LuxHttpErrorInterceptor implements HttpInterceptor {
   static dataStream: ReplaySubject<any> = new ReplaySubject(1);
@@ -21,12 +21,27 @@ export class LuxHttpErrorInterceptor implements HttpInterceptor {
         () => {
           LuxHttpErrorInterceptor.dataStream.next([]);
         },
-        error => {
-          if (error instanceof HttpErrorResponse && error.status === 400) {
-            LuxHttpErrorInterceptor.dataStream.next(error.error.errors);
+        (errorResponse) => {
+          if (errorResponse instanceof HttpErrorResponse && errorResponse.status === 400) {
+            LuxHttpErrorInterceptor.dataStream.next(LuxHttpErrorInterceptor.extractErrors(errorResponse.error));
           }
         }
       )
     );
+  }
+
+  static extractErrors(error: any | null) {
+    let errors;
+    if (typeof error === 'object') {
+      if (error.hasOwnProperty('errors')) {
+        errors = error['errors'];
+      } else if (error.hasOwnProperty('violations')) {
+        errors = error['violations'];
+      } else {
+        errors = [];
+      }
+    }
+
+    return errors;
   }
 }