+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## New
 - **lux-lookup-autocomplete**: Zusätzliches Outputevent "luxDataLoadedAsArray" eingeführt. [Issue 225](https://github.com/IHK-GfI/lux-components/issues/225)
 - **lux-lookup-combobox**: Zusätzliches Outputevent "luxDataLoadedAsArray" eingeführt. [Issue 225](https://github.com/IHK-GfI/lux-components/issues/225)
+- **lux-http-error**: Http-Fehler (Status 400) aus den Properties "errors" und "violations" anzeigen. [Issue 227](https://github.com/IHK-GfI/lux-components/issues/227)
 
 # Version 13.1.0
 ## New